+++ utils.py
@@ -10,8 +10,8 @@ Type          : Python module
 Auteur        : Pierre Théberge
 Compagnie     : Innovations, Performances, Technologies inc.
 Créé le       : 2025-08-05
-Modifié le    : 2026-06-25
-Version       : 0.5.13
+Modifié le    : 2026-07-09
+Version       : 0.5.14
 Copyright     : Pierre Théberge
 
 Description
@@ -87,6 +87,7 @@ Modifications
 0.5.12 - 2026-04-21   [ES-28] : Robustesse : except Exception remplacé par des exceptions
                                spécifiques Selenium et Python dans tous les blocs try/except.
 0.5.13 - 2026-06-25   [ES-27] : Synchronisation de version (aucun changement fonctionnel).
+0.5.14 - 2026-07-09   [CR]    : Synchronisation de version (aucun changement fonctionnel).
 
 Paramètres
 ----------